Enhancement: Better organization of the code with multiple files
Split code into multiple files to make it easier for maintenance.
