+++ gomaxscale.go
@@ -4,97 +4,15 @@ import (
 	"bytes"
 	"crypto/sha1"
 	"encoding/hex"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
-	"log"
 	"net"
-	"regexp"
 	"strconv"
 	"time"
 )
 
 const initBufferSize = 1024
-
-var (
-	reDDLEvent = regexp.MustCompile(`{"namespace":`)
-	reDMLEvent = regexp.MustCompile(`{"domain":`)
-)
-
-// Options stores all available options to connect with MaxScale.
-type Options struct {
-	auth struct {
-		user     string
-		password string
-	}
-	timeouts struct {
-		timeRef func() time.Time
-		read    time.Duration
-		write   time.Duration
-	}
-	uuid       string
-	version    *int
-	gtid       string // Requested GTID position
-	bufferSize int
-	logger     logger
-}
-
-// WithAuth sets the authentication options.
-func WithAuth(user, password string) func(*Options) {
-	return func(o *Options) {
-		o.auth.user = user
-		o.auth.password = password
-	}
-}
-
-// WithGTID sets the GTID position.
-func WithGTID(gtid string) func(*Options) {
-	return func(o *Options) {
-		o.gtid = gtid
-	}
-}
-
-// WithTimeout sets connection timeouts.
-func WithTimeout(readTimeout, writeTimeout time.Duration) func(*Options) {
-	return func(o *Options) {
-		o.timeouts.read = readTimeout
-		o.timeouts.write = writeTimeout
-	}
-}
-
-// WithUUID sets the UUID of the client.
-func WithUUID(uuid string) func(*Options) {
-	return func(o *Options) {
-		o.uuid = uuid
-	}
-}
-
-// WithVersion sets the binlog version to use.
-func WithVersion(version int) func(*Options) {
-	return func(o *Options) {
-		o.version = &version
-	}
-}
-
-// WithBufferSize sets the buffer size for the data stream.
-func WithBufferSize(bufferSize int) func(*Options) {
-	return func(o *Options) {
-		o.bufferSize = bufferSize
-	}
-}
-
-// WithLogger sets the logger to report issues when processing the data stream.
-func WithLogger(logger logger) func(*Options) {
-	return func(o *Options) {
-		o.logger = logger
-	}
-}
-
-type logger interface {
-	Print(v ...interface{})
-	Printf(format string, v ...interface{})
-}
 
 // Consumer is a Go implementation of a MaxScale CDC consumer.
 type Consumer struct {
@@ -109,14 +27,7 @@ type Consumer struct {
 
 // NewConsumer creates a new Consumer instance.
 func NewConsumer(address, database, table string, optFuncs ...func(*Options)) *Consumer {
-	var opts Options
-	opts.uuid = "XXX-YYY_YYY"
-	opts.timeouts.read = 2 * time.Second
-	opts.timeouts.write = 2 * time.Second
-	opts.timeouts.timeRef = time.Now
-	opts.bufferSize = 4096
-	opts.logger = log.Default()
-
+	opts := newDefaultOptions()
 	for _, f := range optFuncs {
 		f(&opts)
 	}
@@ -299,195 +210,4 @@ func (g *Consumer) disconnect(conn net.Conn) {
 
 	close(g.events)
 	close(g.done)
-}
-
-type stream struct {
-	conn        net.Conn
-	readTimeout time.Duration
-	timeRef     func() time.Time
-
-	buffer           bytes.Buffer
-	bufferIdentation int
-	bufferSize       int
-}
-
-func (s *stream) scan() ([]CDCEvent, error) {
-	var responses []bytes.Buffer
-	var loops int
-
-	for {
-		if err := s.conn.SetReadDeadline(s.timeRef().Add(s.readTimeout)); err != nil {
-			return nil, fmt.Errorf("failed to set read timeout: %w", err)
-		}
-
-		buffer := make([]byte, s.bufferSize)
-		n, err := s.conn.Read(buffer)
-		if err != nil {
-			return nil, fmt.Errorf("failed to read response: %w", err)
-		}
-		buffer = buffer[:n]
-
-		var nonJSON bool
-		for _, b := range buffer {
-			var action bool
-			if b == '{' {
-				s.bufferIdentation++
-				action = true
-			} else if b == '}' {
-				s.bufferIdentation--
-				action = true
-			}
-			if s.bufferIdentation == 0 && s.buffer.Len() == 0 {
-				// random test outside of the JSON object is a server error
-				// message being returned
-				nonJSON = true
-			} else if s.bufferIdentation > 0 && nonJSON {
-				// non-JSON data ended and new JSON data started
-				var response bytes.Buffer
-				response.Write(s.buffer.Bytes())
-
-				if len(bytes.TrimSpace(response.Bytes())) > 0 {
-					responses = append(responses, response)
-				}
-
-				s.buffer.Reset()
-				nonJSON = false
-			}
-			if err := s.buffer.WriteByte(b); err != nil {
-				return nil, fmt.Errorf("failed to process response byte: %w", err)
-			}
-			if action && s.bufferIdentation == 0 {
-				var response bytes.Buffer
-				response.Write(s.buffer.Bytes())
-
-				if len(bytes.TrimSpace(response.Bytes())) > 0 {
-					responses = append(responses, response)
-				}
-
-				s.buffer.Reset()
-				action = false
-			}
-		}
-		if nonJSON {
-			// non-JSON data must fit inside a single buffer read call as we
-			// can't determinate when it finishes
-			var response bytes.Buffer
-			response.Write(s.buffer.Bytes())
-
-			if len(bytes.TrimSpace(response.Bytes())) > 0 {
-				responses = append(responses, response)
-			}
-
-			s.buffer.Reset()
-		}
-
-		if len(responses) > 0 {
-			break
-		}
-
-		loops++
-		if loops > 100 {
-			return nil, errors.New("too many network iterations to find a json object")
-		}
-	}
-
-	var events []CDCEvent
-	for i := range responses {
-		if event, err := s.decodeEvent(responses[i].Bytes()); err == nil {
-			events = append(events, event)
-		} else {
-			return nil, err
-		}
-	}
-	return events, nil
-}
-
-func (s *stream) decodeEvent(response []byte) (CDCEvent, error) {
-	switch {
-	case reDDLEvent.Match(response):
-		var ddlEvent DDLEvent
-		if err := json.Unmarshal(response, &ddlEvent); err != nil {
-			return nil, fmt.Errorf("failed to decode ddl event: %w", err)
-		}
-		return ddlEvent, nil
-
-	case reDMLEvent.Match(response):
-		var dmlEvent DMLEvent
-		if err := json.Unmarshal(response, &dmlEvent); err != nil {
-			return nil, fmt.Errorf("failed to decode dml event: %w", err)
-		}
-		dmlEvent.RawData = response
-		return dmlEvent, nil
-	}
-
-	if err := checkResponseError(response); err != nil {
-		return nil, err
-	}
-	return nil, fmt.Errorf("unknown maxscale event type: %s", string(response))
-}
-
-func checkResponseError(response []byte) error {
-	if bytes.Contains(bytes.ToLower(response), []byte("err")) {
-		return fmt.Errorf("error raised from maxscale: %s", string(response))
-	}
-	return nil
-}
-
-// CDCEventType is the type of the event.
-type CDCEventType string
-
-// List of possible event types.
-const (
-	// CDCEventTypeDDL DDL (Data Definition Language) are for database changes.
-	CDCEventTypeDDL CDCEventType = "ddlEvent"
-	// CDCEventTypeDML DML (Data Manipulation Language) are for data changes.
-	CDCEventTypeDML CDCEventType = "dmlEvent"
-)
-
-// CDCEvent is the CDC event received from MaxScale.
-type CDCEvent interface {
-	EventType() CDCEventType
-}
-
-// DDLEvent is a MaxScale DDL event.
-//
-// https://github.com/mariadb-corporation/MaxScale/blob/6.2/Documentation/Routers/KafkaCDC.md#overview
-type DDLEvent struct {
-	Namespace string `json:"namespace"`
-	Type      string `json:"type"`
-	Name      string `json:"name"`
-	Table     string `json:"table"`
-	Database  string `json:"database"`
-	Version   int    `json:"version"`
-	GTID      string `json:"gtid"`
-	Fields    []struct {
-		Name     string      `json:"name"`
-		Type     interface{} `json:"type"`
-		RealType *string     `json:"real_type"`
-		Length   *int        `json:"length"`
-		Unsigned *bool       `json:"unsigned"`
-	} `json:"fields"`
-}
-
-// EventType returns the type of the event.
-func (d DDLEvent) EventType() CDCEventType {
-	return CDCEventTypeDDL
-}
-
-// DMLEvent is a MaxScale DML event.
-//
-// https://github.com/mariadb-corporation/MaxScale/blob/6.2/Documentation/Routers/KafkaCDC.md#overview
-type DMLEvent struct {
-	Domain      int    `json:"domain"`
-	ServerID    int    `json:"server_id"`
-	Sequence    int    `json:"sequence"`
-	EventNumber int    `json:"event_number"`
-	Timestamp   int64  `json:"timestamp"`
-	Type        string `json:"event_type"`
-	RawData     []byte `json:"-"`
-}
-
-// EventType returns the type of the event.
-func (d DMLEvent) EventType() CDCEventType {
-	return CDCEventTypeDML
 }
